ULTIMA4: Moving mouse cursor code into Screen class
